+++ R/check-cran.r
@@ -56,8 +56,9 @@ check_cran <- function(pkgs, libpath = file.path(tempdir(), "R-lib"),
     }
     available_src <- available_packages(repos, "source")
 
-    message("Determining available packages") # ----------------------------------
+    message("Determining packages to update... ", appendLF = FALSE) # ------------
     deps <- package_deps(pkgs, repos = repos, type = type, dependencies = TRUE)
+    message(paste(deps$package, collapse = ", "))
     update(deps, Ncpus = threads, quiet = TRUE)
 
     message("Downloading source packages for checking") #-------------------------