+++ R/install-remote.R
@@ -43,8 +43,20 @@ install_remote <- function(remote, ..., force = FALSE, quiet = FALSE) {
   install(source, ..., quiet = quiet, metadata = metadata)
 }
 
+try_install_remote <- function(..., quiet) {
+  tryCatch(
+    install_remote(..., quiet = quiet),
+    error = function(e) {
+      if (!quiet) {
+        message("Installation failed: ", conditionMessage(e))
+      }
+      FALSE
+    }
+  )
+}
+
 install_remotes <- function(remotes, ...) {
-  invisible(vapply(remotes, install_remote, ..., FUN.VALUE = logical(1)))
+  invisible(vapply(remotes, try_install_remote, ..., FUN.VALUE = logical(1)))
 }
 
 # Add metadata